infra/requirements.txt
aws-cdk-lib==2.205.0
constructs>=10.0.0,<11.0.0
aws-cdk.aws-apigatewayv2-alpha>=2.0.0a0,<3.0.0
aws-cdk.aws-apigatewayv2-integrations-alpha>=2.0.0a0,<3.0.0
aws-cdk.aws-lambda-python-alpha>=2.0.0a0,<3.0.0

